+++ src/ui-lib/buttons.tsx
@@ -1,12 +1,11 @@
 import React, { FC, MouseEventHandler } from 'react';
 import styled from 'styled-components';
 import {
-  TFontProperties, TButtonStyle, TButtonTextStyle, TColorSet,
+  TFontProperties, TButtonStyle, TButtonTextStyle, TColorSet, TButtonProps,
 } from './styles.types';
-import { PlusIcon, BasketIcon, MinusIcon } from './index';
-import DeletePic from '../assets/images/icons/trash-icon.svg';
-// import EditPic from '../assets/images/icons/edit-icon.svg';
-import { ReactComponent as EditPic } from '../assets/images/icons/edit-icon.svg';
+import {PlusIcon, BasketIcon, MinusIcon, EditIcon} from './index';
+
+
 
 const buttonFont : TFontProperties = {
   family: 'Alagreya Sans',
@@ -23,26 +22,7 @@ const blueButtonStyles : TButtonStyle = {
   fontColor: '#FFFFFF',
   fontProperties: buttonFont,
 };
-const whiteIconStyle: TColorSet = {
-  defaultColor: '#FFFFFF',
-  hoverColor: '#FFFFFF',
-  activeColor: '#FFFFFF',
-  disabledColor: '#CCCCCC',
-};
 
-const redIconStyle: TColorSet = {
-  defaultColor: '#FF413B',
-  hoverColor: '#E53B35',
-  activeColor: '#CC342F',
-  disabledColor: '#CCCCCC',
-};
-
-const blueIconStyle: TColorSet = {
-  defaultColor: '#008AFF',
-  hoverColor: '#007CE5',
-  activeColor: '#006ECC',
-  disabledColor: '#CCCCCC',
-};
 
 const redButtonStyles : TButtonStyle = {
   defaultColor: '#FF413B',
@@ -54,116 +34,11 @@ const redButtonStyles : TButtonStyle = {
 };
 
 const iconDistance = 8;
-const getColor = (disabled : boolean, baseColor : string, disabledColor: string) : string => {
-  if (disabled) {
-    return disabledColor;
-  }
-  return baseColor;
-};
+
 interface IButtonStylesInterface {
   buttonStyle: TButtonStyle;
+  disabled?: boolean;
 }
-
-type TBasicIconProps = {
-  defaultColor: string;
-  hoverColor: string;
-  activeColor: string;
-};
-const basicIconStyles = `
-  width: 24px;
-  height: 24px;
-  display: block;
-`;
-
-const BasicEditIcon = styled<TBasicIconProps>(EditPic)`
-  width: 24px;
-  height: 24px;
-  display: block;
-  & > svg > path {
-    stroke: ${({ defaultColor }) => defaultColor};
-    }
-  &:hover > svg > path {
-    stroke: ${({ hoverColor }) => hoverColor};
-    }
-  &:focus > svg > path {
-    stroke: ${({ defaultColor }) => defaultColor};
-    }
-  &:active > svg > path {
-    stroke: ${({ activeColor }) => activeColor};
-    }
-`;
-
-type TIconProps = {
-  onClick?: MouseEventHandler<HTMLImageElement>;
-  disabled: boolean;
-  colors: TColorSet;
-};
-/* const DeleteIcon : FC<TIconProps> = ({ onClick = undefined, disabled, colors }) => {
-  const {
-    defaultColor,
-    activeColor,
-    hoverColor,
-    disabledColor,
-  } = colors;
-  const defColor = (disabledState : boolean) : string => {
-    if (disabledState) {
-      return disabledColor;
-    }
-    return defaultColor;
-  };
-  const hovColor = (disabledState : boolean) : string => {
-    if (disabledState) {
-      return disabledColor;
-    }
-    return hoverColor;
-  };
-  const actColor = (disabledState : boolean) : string => {
-    if (disabledState) {
-      return disabledColor;
-    }
-    return activeColor;
-  };
-  return (
-    <BasicIcon
-      src={DeletePic as string}
-      alt='Иконка удаления'
-      onClick={onClick}
-      defaultColor={defColor(disabled)}
-      activeColor={actColor(disabled)}
-      hoverColor={hovColor(disabled)} />
-  );
-};
-*/
-const EditIcon : FC<TIconProps> = ({ onClick = undefined, disabled = false, colors }) => {
-  const {
-    defaultColor,
-    hoverColor,
-    activeColor,
-    disabledColor,
-  } = colors;
-  return (
-    <BasicEditIcon
-      defaultColor={getColor(disabled, defaultColor, disabledColor)}
-      hoverColor={getColor(disabled, hoverColor, disabledColor)}
-      activeColor={getColor(disabled, activeColor, disabledColor)} />
-  );
-};
-
-/* const EditIconRaw = styled<TRawIconProps>(EditPic)`
-  ${basicIconStyles}
-  & > path {
-    stroke: ${({ colors: { defaultColor } }) => defaultColor};
-    }
-  &:hover > path {
-    stroke: ${({ colors: { hoverColor } }) => hoverColor};
-    }
-  &:focus > path {
-    stroke: ${({ colors: { defaultColor } }) => defaultColor};
-    }
-  &:active > path {
-    stroke: ${({ colors: { activeColor } }) => activeColor};
-    }
-`; */
 
 const BasicNormalButton = styled.button<IButtonStylesInterface>`
   padding: 8px 16px;
@@ -172,7 +47,7 @@ const BasicNormalButton = styled.button<IButtonStylesInterface>`
   font-size: ${({ buttonStyle: { fontProperties: { size } } }) => size} px;
   font-weight: ${({ buttonStyle: { fontProperties: { weight } } }) => weight};
   line-height: ${({ buttonStyle: { fontProperties: { height } } }) => height} px;
-  background-color: ${({ buttonStyle: { defaultColor } }) => defaultColor};
+  background-color: ${({ buttonStyle: { defaultColor, disabledColor }, disabled }) => defaultColor};
   color: ${({ buttonStyle: { fontColor } }) => fontColor};
   display: flex;
   flex-flow: row nowrap;
@@ -222,26 +97,25 @@ const ButtonText = styled.p<TButtonTextStyle>`
   margin-block-end: 0;
   margin-inline-start: 0;
   margin-inline-end: 0;
-  // padding-left: ${({ paddingLeft = 0 }) => paddingLeft}px;
 `;
 
-export const EditPostButton = (onClick: MouseEventHandler<HTMLButtonElement>) => (
+export const EditPostButton : FC<TButtonProps> = (
+  onClick, disabled) => (
   <BasicNormalButton buttonStyle={blueButtonStyles} onClick={onClick}>
-    <EditIcon colors={whiteIconStyle} disabled={false} />
+    <EditIcon />
     <ButtonText paddingLeft={iconDistance}>Редактировать запись</ButtonText>
   </BasicNormalButton>
 );
 
 export const DeletePostButton = (
   onClick: MouseEventHandler<HTMLButtonElement>,
-  disabled : false = false,
 ) => (
   <BasicInvertedButton buttonStyle={redButtonStyles} onClick={onClick}>
-    <DeleteIcon disabled={disabled} colors={redButtonStyles} />
+    <BasketIcon  />
     <ButtonText paddingLeft={iconDistance}>Удалить запись</ButtonText>
   </BasicInvertedButton>
 );
-/*
+
 export const SavePostButton = (onClick: MouseEventHandler<HTMLButtonElement>) => (
   <BasicNormalButton buttonStyle={blueButtonStyles} onClick={onClick}>
     Сохранить запись
